Uploading some files
These files are examples of some of the work I did in my final statistics course at Roanoke College. They are all written in R, and for each project I've included the R file that I (or we in some cases) used as scrathwork, the R Markdown file and the HTML file it produced.